more tests

+++ src/test/java/com/samourai/efficiencyscore/processor/TxProcessorTest.java
@@ -14,15 +14,15 @@ public class TxProcessorTest {
     @Test
     public void testCheckCoinjoinPattern() {
         // 8e56317360a548e8ef28ec475878ef70d1371bee3526c017ac22ad61ae5740b8
-        Map<String, Integer> outs = new HashMap<>();
-        outs.put("18JNSFk8eRZcM8RdqLDSgCiipgnfAYsFef", 9850000);
-        outs.put("1PA1eHufj8axDWEbYfPtL8HXfA66gTFsFc", 1270000);
-        outs.put("1JR3x2xNfeFicqJcvzz1gkEhHEewJBb5Zb", 100000);
-        outs.put("1ALKUqxRb2MeFqomLCqeYwDZK6FvLNnP3H", 100000);
+        Map<String, Long> outs = new HashMap<>();
+        outs.put("18JNSFk8eRZcM8RdqLDSgCiipgnfAYsFef", 9850000L);
+        outs.put("1PA1eHufj8axDWEbYfPtL8HXfA66gTFsFc", 1270000L);
+        outs.put("1JR3x2xNfeFicqJcvzz1gkEhHEewJBb5Zb", 100000L);
+        outs.put("1ALKUqxRb2MeFqomLCqeYwDZK6FvLNnP3H", 100000L);
         processCheckCoinjoinPattern(outs, 2, new CoinjoinPattern(2, 100000));
     }
 
-    private void processCheckCoinjoinPattern(Map<String, Integer> outs, int maxNbEntities, CoinjoinPattern expected) {
+    private void processCheckCoinjoinPattern(Map<String, Long> outs, int maxNbEntities, CoinjoinPattern expected) {
         CoinjoinPattern result = txProcessor.checkCoinjoinPattern(outs, maxNbEntities);
 
         Assert.assertEquals(expected.getNbPtcpts(), result.getNbPtcpts());
